drm/msm/sde: cache connector list in sde crtc
Determine which connectors are associated with the crtc during
the beginning of the commit request and reuse the list for the
rest of the commit. This is needed because the drm atomic states
only record connectors/objects that have been updated in the
current commit, rather than all associated plane/connector/etc.
objects.

Change-Id: I51f2379fc574033b4f87e01eccbb3ee8aef31bb3
Signed-off-by: Clarence Ip <cip@codeaurora.org>