drm/msm/sde: remove bandwdith ctrl handling
The power handles added API to add refcouting if
any client is enabled or not for bus bandwidth
ctrl and then send the vote to bus driver. This
should be handled through core perf bandwidth
calculation instead of separate API to enable/disable.
This patch simplifies the logic and allows the
core perf client to send the vote when it is ready.

Change-Id: Ie2fbc8eaf378630fff0ca1629f82caa179487b6f
Signed-off-by: Dhaval Patel <pdhaval@codeaurora.org>