drm/msm/dp: add support for HDCP 2.2
Add a module to control the interactions between the Receiver,
Transmitter, and HDCP enforcing hardware layer as part of the
HDCP 2.2 protocol. This module is used by DisplayPort interface
driver to initiate HDCP 2.2 with a sink and provide a secure
link for content.

This implementation is based on a snapshot of the DisplayPort
HDCP 2.2 module as per this commit Ief0a8ce77a932 ("msm: mdss:
hdcp2p2: fix sysfs node name") on kernel-4.4.

CRs-Fixed: 2057884
Change-Id: Ic1f493c78afc9275ad62f9494bb5b3f26d1dab70
Signed-off-by: Tatenda Chipeperekwa <tatendac@codeaurora.org>